mwCheckbox {manipulateWidget}	R Documentation

Add a checkbox to a manipulateWidget gadget

Description

Add a checkbox to a manipulateWidget gadget

Usage

mwCheckbox(value = FALSE, label = NULL, ..., .display = TRUE)

Arguments

value
Initial value of the input.

label
Display label for the control. If NULL, the name of the corresponding variable is used.

...
Other arguments passed to functioncheckboxInput

.display
expression that evaluates to TRUE or FALSE, indicating when the input control should be shown/hidden.

Value

A function that will generate the input control.

See Also

Other controls: mwCheckboxGroup(), mwDateRange(), mwDate(), mwGroup(), mwNumeric(), mwPassword(), mwRadio(), mwSelectize(), mwSelect(), mwSharedValue(), mwSlider(), mwText()

Examples

if(require(plotly)) {
 manipulateWidget(
   {
       plot_ly(iris, x = ~Sepal.Length, y = ~Sepal.Width,
               color = ~Species, type = "scatter", mode = "markers") %>%
         layout(showlegend = legend)
   },
   legend = mwCheckbox(TRUE, "Show legend")
 )
}

[Package manipulateWidget version 0.10.1 Index]