mwDate {manipulateWidget}	R Documentation

Add a date picker to a manipulateWidget gadget

Description

Add a date picker to a manipulateWidget gadget

Usage

mwDate(value = NULL, label = NULL, ..., .display = TRUE)

Arguments

value
Default value of the input.

label
Display label for the control. If NULL, the name of the corresponding variable is used.

...
Other arguments passed to functiondateInput

.display
expression that evaluates to TRUE or FALSE, indicating when the input control should be shown/hidden.

Value

A function that will generate the input control.

See Also

Other controls: mwCheckboxGroup(), mwCheckbox(), mwDateRange(), mwGroup(), mwNumeric(), mwPassword(), mwRadio(), mwSelectize(), mwSelect(), mwSharedValue(), mwSlider(), mwText()

Examples

if (require(dygraphs) && require(xts)) {
  mydata <- xts(rnorm(365), order.by = as.Date("2017-01-01") + 0:364)

  manipulateWidget(
    dygraph(mydata) %>% dyEvent(date, "Your birthday"),
    date = mwDate("2017-03-27", label = "Your birthday date",
                  min = "2017-01-01", max = "2017-12-31")
  )
}

[Package manipulateWidget version 0.10.1 Index]