mwGroup {manipulateWidget}	R Documentation

Group inputs in a collapsible box

Description

This function generates a collapsible box containing inputs. It can be useful when there are a lot of inputs and one wants to group them.

Usage

mwGroup(..., label = NULL, .display = TRUE)

Arguments

...
inputs that will be grouped in the box

label
label of the group inputs

.display
expression that evaluates to TRUE or FALSE, indicating when the group should be shown/hidden.

Value

Input of type "group".

See Also

Other controls: mwCheckboxGroup(), mwCheckbox(), mwDateRange(), mwDate(), mwNumeric(), mwPassword(), mwRadio(), mwSelectize(), mwSelect(), mwSharedValue(), mwSlider(), mwText()

Examples

if(require(dygraphs)) {
  mydata <- data.frame(x = 1:100, y = rnorm(100))
  manipulateWidget(
    dygraph(mydata[range[1]:range[2], ],
            main = title, xlab = xlab, ylab = ylab),
    range = mwSlider(1, 100, c(1, 100)),
    "Graphical parameters" = mwGroup(
      title = mwText("Fictive time series"),
      xlab = mwText("X axis label"),
      ylab = mwText("Y axis label")
    )
  )
}

[Package manipulateWidget version 0.10.1 Index]