Generate the UI of a manipulateWidget gadget

This function can be used if you desire to create a gadget that has the same UI as a manipulateWidget gadget but with a custom server logic.

mwUI(..., .controlPos = c("left", "top", "right", "bottom", "tab"),
  .tabColumns = 2, .updateBtn = FALSE, .main = "", .outputFun = NULL,
  .outputId = "output", .titleBar = TRUE, .updateInputs = NULL,
  .compare = NULL, .compareLayout = c("v", "h"), .controlList = NULL,
  .container = miniUI::miniContentPanel, .style = "",
  .env = parent.frame())

Arguments

...

One or more named control arguments created with functions mwSlider, mwText, etc. The name of each control is the name of the variable the controls modifies in the expression. One can also create a group of inputs by passing a list of such control arguments. for instance mygroup = list(txt = mwText(""), nb = mwNumeric(0)) creates a group of inputs named mygroup with two inputs named "txt" and "nb".

.controlPos

Where controls should be placed ? By default, they are placed in the left, next to the graphic. If controlPos = "tab", two tabs are created: one containing controls and the other containing the graphic.

.tabColumns

If controls are placed in a distinct tab, how many columns should be used ? This parameter is used only if controlPos = "tab"

.updateBtn

Should an update button be added to the controls ? If TRUE, then the graphic is updated only when the user clicks on the update button.

.main

Title of the shiny gadget

.outputFun

The output function for the desired htmlwidget.

.outputId

Id of the output element in the shiny interface.

.titleBar

Whether to include a title bar with controls in the widget

.updateInputs

This parameter is similar to `.display` and can be used to dynamically update input controls. It must be a named list where names correspond to names of input controls and values are named lists of expressions where names correspond to arguments of the input generator function.

.compare

Sometimes one wants to compare the same chart but with two different sets of parameters. This is the purpose of this argument. It must be a named list whose names are the names of the inputs that should vary between the two charts. All other parameters are common to the two charts and changing their values will change the two charts. Each element of the list must be a vector or a list of length 2 with the initial values of the corresponding parameter for each chart. It can also be NULL. In this case, the parameter is initialized with the default value for the two charts.

.compareLayout

Used only when .compare is set. Possible values are "v" for vertical layout (one chart above the other) and "h" for horizontal layout (one chart on the right of the other)

.controlList

List of input controls. This is an alternative to specifying directly the controls through the ... arguments.

.container

tag function that will be used to enclose the UI.

.style

CSS style to apply to the container element.

.env

Environment used to evaluate the inital values of controls. This parameter may have an impact on the result only when .updateInputs is used.

Value

A shiny.tag.list object that can be used in function runGadget as ui parameter.