Generate the UI of a manipulateWidget gadget

This function can be used if you desire to create a gadget that has the same UI as a manipulateWidget gadget but with a custom server logic.

mwUI(..., .controlPos = c("left", "top", "right", "bottom", "tab"),
  .tabColumns = 2, .updateBtn = FALSE, .main = "",
  .content = htmlOutput("output", style = "height:100%;width:100%"))

Arguments

...

One or more named control arguments created with functions mwSlider, mwText, etc. The name of each control is the name of the variable the controls modifies in the expression. One can also create a group of inputs by passing a list of such control arguments. for instance mygroup = list(txt = mwText(""), nb = mwNumeric(0)) creates a group of inputs named mygroup with two inputs named "txt" and "nb".

.controlPos

Where controls should be placed ? By default, they are placed in the left, next to the graphic. If controlPos = "tab", two tabs are created: one containing controls and the other containing the graphic.

.tabColumns

If controls are placed in a distinct tab, how many columns should be used ? This parameter is used only if controlPos = "tab"

.updateBtn

Should an update button be added to the controls ? If TRUE, then the graphic is updated only when the user clicks on the update button.

.main

Title of the shiny gadget

.content

HTML of the main content of the application, where outputs are displayed.

Value

A shiny.tag.list object that can be used in function runGadget as ui parameter.